Split the app.php into several files for better readibility, added some comments, fixed Dockerfile, added export from database for quicker setup
