+++ app/classes/articleFactory.php
@@ -1,0 +1,9 @@
+<?php
+class ArticleFactory
+{
+    public function createArticle($id, $name, $text, $publicationDate, $contributorName)
+    {
+        return new Article($name, $text, $publicationDate, $contributorName, $id);
+    }
+}
+?>
